+++ app/collections/collectionHandle/page.tsx
@@ -1,3 +1,0 @@
-export default function Collection() {
-  return <>HERE IS THE COLLECTION HANDLE PAGE</>;
-}