+++ lib/shopify/queries/cart.ts
@@ -1,0 +1,10 @@
+import { cartFragment } from "./fragments";
+
+export const getCartQuery = `#graphql
+  query getCart($cartId: ID!) {
+    cart(id: $cartId) {
+      ...cart
+    }
+  }
+  ${cartFragment}
+`;